Move intrinsic_gas to protocol
