+++ silkworm/core/protocol/intrinsic_gas_test.cpp
@@ -18,9 +18,9 @@
 
 #include <catch2/catch.hpp>
 
-#include <silkworm/core/protocol/param.hpp>
+#include "param.hpp"
 
-namespace silkworm {
+namespace silkworm::protocol {
 
 TEST_CASE("EIP-2930 intrinsic gas") {
     std::vector<AccessListEntry> access_list{
@@ -49,7 +49,7 @@ TEST_CASE("EIP-2930 intrinsic gas") {
     };
 
     intx::uint128 g0{intrinsic_gas(txn, EVMC_ISTANBUL)};
-    CHECK(g0 == protocol::fee::kGTransaction + 2 * protocol::fee::kAccessListAddressCost + 2 * protocol::fee::kAccessListStorageKeyCost);
+    CHECK(g0 == fee::kGTransaction + 2 * fee::kAccessListAddressCost + 2 * fee::kAccessListStorageKeyCost);
 }
 
-}  // namespace silkworm
+}  // namespace silkworm::protocol
