+++ silkworm/silkrpc/core/evm_executor.cpp
@@ -26,8 +26,8 @@
 #include <evmc/evmc.hpp>
 #include <intx/intx.hpp>
 
-#include <silkworm/core/chain/intrinsic_gas.hpp>
 #include <silkworm/core/common/util.hpp>
+#include <silkworm/core/protocol/intrinsic_gas.hpp>
 #include <silkworm/core/protocol/param.hpp>
 #include <silkworm/silkrpc/common/log.hpp>
 #include <silkworm/silkrpc/common/util.hpp>
@@ -227,7 +227,7 @@ boost::asio::awaitable<ExecutionResult> EVMExecutor::call(
 
                 const evmc_revision rev{evm.revision()};
                 const intx::uint256 base_fee_per_gas{evm.block().header.base_fee_per_gas.value_or(0)};
-                const intx::uint128 g0{intrinsic_gas(txn, rev)};
+                const intx::uint128 g0{protocol::intrinsic_gas(txn, rev)};
                 SILKWORM_ASSERT(g0 <= UINT64_MAX);  // true due to the precondition (transaction must be valid)
 
                 const auto error = pre_check(evm, txn, base_fee_per_gas, g0);
